Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main()
-{int a;
-scanf("%d",&a);
-int b;
-scanf("%d",&b);
+{float a;
+scanf("%f",&a);
+float b;
+scanf("%f",&b);
 // int res;
 // res=a*b;
 printf("Product:%d*%d",a,b);
